+++ packages/multichain-api-middleware/src/index.ts
@@ -2,6 +2,6 @@ export { walletGetSession } from './handlers/wallet-getSession';
 export { walletInvokeMethod } from './handlers/wallet-invokeMethod';
 export { walletRevokeSession } from './handlers/wallet-revokeSession';
 
-export { multichainMethodCallValidatorMiddleware } from './middlewares/multichainMethodCallValidator';
+export { multichainMethodCallValidatorMiddleware } from './middlewares/multichainMethodCallValidatorMiddleware';
 export { MultichainMiddlewareManager } from './middlewares/MultichainMiddlewareManager';
 export { MultichainSubscriptionManager } from './middlewares/MultichainSubscriptionManager';
